requirements1.txt
asgiref==3.7.2
Django==3.0
django-livereload-server==0.5.1
djongo==1.3.6
dnspython==2.5.0
mysqlclient==2.2.1
pymongo==4.6.1
python-decouple==3.8
pytz==2023.3.post1
setuptools==68.2.2
sqlparse==0.2.4
tornado==6.4
wheel==0.41.2
